backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-httpauth
python-dotenv
waitress